+++ Editor/Modules/FogModeAnalyzer.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using Unity.ProjectAuditor.Editor.Core;
 using Unity.ProjectAuditor.Editor.Diagnostic;
 using UnityEditor;
@@ -6,19 +7,29 @@ using UnityEngine.Rendering;
 
 namespace Unity.ProjectAuditor.Editor.Modules
 {
+    public enum FogMode
+    {
+        Linear,
+        Exponential,
+        ExponentialSquarred,
+        Automatic
+    }
+
     class FogModeAnalyzer : ISettingsModuleAnalyzer
     {
         private static readonly Descriptor k_FogModeDescriptor = new Descriptor(
             "PAS1003",
-            "Graphics: Fog Mode",
+            "Graphics: Fog Shader Variant Stripping",
             new[] {Area.BuildSize},
-            "Enabling Fog Stripping will result in additional shader variants, thus increasing build size.",
-            "To reduce the build size, switch Fog Modes at <b>Edit ➔ Project Settings ➔ Graphics ➔ Fog Modes</b> to <b>Automatic</b>.")
+            "FogMode shader variants are always built. Forcing Fog shader variants to be built can increase the build size.",
+            "To reduce the number of shader variants, change <b>Edit ➔ Project Settings ➔ Graphics ➔ Fog Modes</b> to <b>Automatic</b> or disable <b>Linear/Exponential/Exponential Squared</b>.")
         {
             fixer = (issue =>
             {
                 RemoveFogStripping();
-            })
+            }),
+
+            messageFormat = "Graphics: FogMode {0} shader variants are always included in the build."
         };
 
         public void Initialize(ProjectAuditorModule module)
@@ -28,27 +39,56 @@ namespace Unity.ProjectAuditor.Editor.Modules
 
         public IEnumerable<ProjectIssue> Analyze(ProjectAuditorParams projectAuditorParams)
         {
-            if (IsFogStrippingCustom())
+            if (!IsFogStrippingEnabled(FogMode.Automatic))
             {
-                yield return ProjectIssue.Create(IssueCategory.ProjectSetting, k_FogModeDescriptor)
+                yield return ProjectIssue.Create(IssueCategory.ProjectSetting, k_FogModeDescriptor, GetFogModesEnabledString())
                     .WithLocation("Project/Graphics");
             }
         }
 
-        internal static bool IsFogStrippingCustom()
+        static string GetFogModesEnabledString()
         {
             var serializedObject = new SerializedObject(GraphicsSettings.GetGraphicsSettings());
-            var mode = serializedObject.FindProperty("m_FogStripping").enumValueIndex;
+            string message = "";
 
-            //As we can't access the enum from here, we can't cast it and check the value
-            //1 is for "Custom" - 0 for "Automatic"
-            if (mode == 1)
+            var linearFog = serializedObject.FindProperty("m_FogKeepLinear").boolValue;
+            var expFog = serializedObject.FindProperty("m_FogKeepExp").boolValue;
+            var exp2Fog = serializedObject.FindProperty("m_FogKeepExp2").boolValue;
+
+            if (linearFog) message += "- Linear ";
+            if (expFog) message += "- Exponential ";
+            if (exp2Fog) message += "- Exponential Squarred ";
+
+            return message;
+        }
+
+        internal static bool IsFogStrippingEnabled(FogMode fogMode)
+        {
+            var serializedObject = new SerializedObject(GraphicsSettings.GetGraphicsSettings());
+            bool isEnabled = false;
+
+            switch (fogMode)
             {
-                return true;
+                case FogMode.Automatic:
+                    isEnabled = serializedObject.FindProperty("m_FogStripping").enumValueIndex == 0; //Automatic mode
+                    break;
+
+                case FogMode.Exponential:
+                    isEnabled = serializedObject.FindProperty("m_FogKeepExp").boolValue;
+                    break;
+
+                case FogMode.ExponentialSquarred:
+                    isEnabled = serializedObject.FindProperty("m_FogKeepExp2").boolValue;
+                    break;
+
+                case FogMode.Linear:
+                    isEnabled = serializedObject.FindProperty("m_FogKeepLinear").boolValue;
+                    break;
             }
 
-            return false;
+            return isEnabled;
         }
+
 
         internal static void RemoveFogStripping()
         {